Revert "Feature/weightage"
