+++ src/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route,Navigate } from "react-router-dom";
 import { AuthContextProvider } from "./context/AuthContext.jsx";
 import { LOGIN,APPLICATIONS,
    DASHBOARD,FORGOTPASSWORD,
-   CHANGEPASSWORD,TASK,ADDTASK,ADDQUIZ,INDEX,WEIGHTAGE } from "./routes/routes.jsx";
+   CHANGEPASSWORD,TASK,ADDTASK,ADDQUIZ,INDEX } from "./routes/routes.jsx";
 //Define your routes for APP here
 import LoginPage from "./pages/login/LoginPage.jsx";
 import ChangePassword from "./pages/changePasswordPage/ResetPasswordPage";
@@ -14,7 +14,6 @@ import TaskModule from "./pages/taskModule/TaskModule.jsx";
 
 import AddTask from "./pages/addTaskPage/AddTaskPage.jsx";
 import AddQuiz from "./pages/addQuiz/AddQuizz.jsx";
-import Weightage from "./pages/addTaskPage/weightage/WeightAge.jsx";
 import ErrorPage from "./pages/errorPage/ErrorPage.jsx";
 //Private Routes will be wrapped in below component
 import PrivateRoute from "./routes/PrivateRoute";
@@ -49,9 +48,7 @@ const App = () => {
              <Route path={ADDQUIZ} element={<AddQuiz />} />
           </Route>
 
-          <Route path={WEIGHTAGE} element={<PrivateRoute />}>
-             <Route path={WEIGHTAGE} element={<Weightage/>} />
-          </Route>
+
 
         <Route path='*' element={<ErrorPage/>} />
         </Routes>
