--- src/pages/addTaskPage/weightage/WeightAge.css
@@ -1,122 +1,0 @@
-
-
-p.icon__title {
-    display: flex;
-    align-items: center;
-    color: #223F64;
-  
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 150%;
-    /* 24px */
-  
-  }
-  
-  .icon__title span {
-    font-size: 24px;
-  }
-  
-  .weightage__header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  
-  .close span {
-    width: 32px;
-    height: 32px;
-  }
-  
-  .input__label {
-    color: #223F64;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 150%;
-    /* 24px */
-  }
-  
-  .input__description {
-    display: flex;
-    width: 400px;
-    height: 56px;
-    padding: 8px 16px;
-    align-items: center;
-    gap: 10px;
-    border-radius: 2px;
-    border: 1px solid #ECECEC;
-    background: var(--base-white, #FFF);
-    outline: 0;
-  }
-  
-  .input__value {
-    display: flex;
-    width: 400px;
-    height: 56px;
-    padding: 8px 16px;
-    align-items: center;
-    gap: 10px;
-    border-radius: 2px;
-    border: 1px solid #ECECEC;
-    background: var(--base-white, #FFF);
-    outline: 0;
-  }
-  
-  /* .des\&value {
-    display: flex;
-    justify-content: space-around;
-    padding-top: 25px;
-  } */
-  .des\&value {
-    display: flex;
-   
-    justify-content: flex-start;
-    padding-top: 25px;
-    gap: 82px;
-  }
-  .btns {
-    max-width: 300px;
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-  
-  
-  
-  }
-  
-  .btns .save__new button {
-    padding: 15px 35px;
-    background: transparent;
-    border: 2px solid #C8C7C7;
-    color: #C8C7C7;
-    border-radius: 15px;
-    cursor: pointer;
-  }
-  
-  
-  .btns .save button {
-    padding: 15px 50px;
-    background: transparent;
-    border: 2px solid #3F88D8;
-    border-radius: 15px;
-    color: var(--secondary-colour, #3F88D8);
-    font-size: 16px;
-    cursor: pointer;
-  
-  }
-  
-  .all_btn {
-    display: flex;
-    justify-content: end;
-    padding-top: 30px;
-    max-width: 960px;
-    width: 100%;
-  }
-  
-  .cancel {
-    color: #e00d0d;
-    font-size: 16px;
-    cursor: pointer;
-  }
-
