+++ src/pages/dashBoard/index.css
@@ -1,15 +1,3 @@
-
-
-.add-container{
-  display: flex;
-  gap: 10px;
-  flex-wrap: wrap;
-}
-.added-batches a {
-  text-decoration: none;
-  color: #A82700;
-}
-
 .greeting {
   background-color: white;
   display: flex;
@@ -55,6 +43,7 @@
 .add__batch {
   padding-top: 20px;
 }
+
 
 
 .batches__list {
