+++ src/routes/routes.jsx
@@ -10,8 +10,6 @@ const DASHBOARD = "/dashboard";
 const APPLICATIONS = "/batch/:id/applications"
 const TASK = "/batch/:id/module"
 const ADDTASK = "/batch/:id/add/task"
-const WEIGHTAGE = "/batch/:id/add/task/weightage"
 const ADDQUIZ = "/batch/:id/add/quiz"
-
-export { DASHBOARD,LOGIN,APPLICATIONS, FORGOTPASSWORD,CHANGEPASSWORD,TASK ,ADDTASK,ADDQUIZ,INDEX,WEIGHTAGE};
+export { DASHBOARD,LOGIN,APPLICATIONS, FORGOTPASSWORD,CHANGEPASSWORD,TASK ,ADDTASK,ADDQUIZ,INDEX};
